Dashboard chart data complete

+++ app/Http/Controllers/RoleController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Role;
 use Illuminate\Http\Request;
+use App\Http\Resources\RoleResource;
+use Illuminate\Support\Facades\Gate;
 use Symfony\Component\HttpFoundation\Response;
 
 class RoleController extends Controller
@@ -15,7 +17,8 @@ class RoleController extends Controller
      */
     public function index()
     {
-        return Role::all();
+        Gate::authorize('view', 'roles');
+        return RoleResource::collection(Role::all());
     }
 
     /**
@@ -26,11 +29,15 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('edit', 'roles');
         $request->validate(['name'=>['required','unique:roles,name']]);
 
         $role = Role::create($request->only('name'));
+        if($permissoins =  $request->permissions){
+            $role->permissions()->sync($permissoins);
+        }
 
-        return response($role, Response::HTTP_CREATED);
+        return response(new RoleResource($role), Response::HTTP_CREATED);
     }
 
     /**
@@ -41,7 +48,8 @@ class RoleController extends Controller
      */
     public function show(Role $role)
     {
-        return response($role, Response::HTTP_OK);
+        Gate::authorize('view', 'roles');
+        return new RoleResource($role);
     }
 
     /**
@@ -53,11 +61,15 @@ class RoleController extends Controller
      */
     public function update(Request $request, Role $role)
     {
+        Gate::authorize('edit', 'roles');
         $request->validate(['name'=>['required','unique:roles,name,'.$role->id]]);
 
         $role->update($request->only('name'));
+        if($permissoins =  $request->permissions){
+            $role->permissions()->sync($permissoins);
+        }
 
-        return response($role, Response::HTTP_ACCEPTED);
+        return response(new RoleResource($role), Response::HTTP_ACCEPTED);
     }
 
     /**
@@ -68,6 +80,8 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        Gate::authorize('edit', 'roles');
+        $role->permissions()->detach();
         $role->delete();
 
         return response(null, Response::HTTP_NO_CONTENT);
